Updating platform os names to reflect latest changes by @imartinezortiz affecting the release process.
