+++ build-tools/generators/src/main/resources/appdata/release.json
@@ -30,10 +30,8 @@
         },
 
         "os":{
-            "type": "string",
-            "description": "The installer version: win32 | win64 | mac, etc.",
-            "enum": ["multiplatform", "win32", "win64", "macosx", "linux-i386", "linux-amd64"],
-            "default": "multiplatform"
+            "$ref": "os.json",
+            "description": "The installer version: win32 | win64 | mac, etc."
         },
         "engineLibPath":{
             "type": "string",
